Make test separate function
