+++ tests/test_ready_valid.py
@@ -17,5 +17,4 @@ def test_basic_ready_valid_sequence():
 
     I = [BitVector.random(8) for _ in range(8)]
     O = [i + 1 for i in I]
-    tester = f.ReadyValidTester(Main, {"I": I, "O": O})
-    tester.compile_and_run("verilator")
+    f.run_ready_valid_test(Main, {"I": I, "O": O}, "verilator")
